requirements.txt
flask
prometheus_client
stix2
pysmile
pyyaml
